+++ src/main/kotlin/euler/Task07_10001stPrime.kt
@@ -7,7 +7,7 @@ object Task07_10001stPrime : EulerTask {
 
     override val name: String = "10001st prime"
 
-    override fun solution(): Int {
+    override fun solution(): Long {
         return findPrimes().take(10_001).last()
     }
